backend/requirements.txt
fastapi
uvicorn
pillow
torch
torchvision
opencv-python
onnxruntime
python-multipart
